Refactoring HTTP handling and unified tasks startup
- move HTTP handlers to http_server module
- add start_http_server function
--> `server.rs` is now maybe redundant?
